Initial Upload | Incomplete but doesnt crash
# Bruce Fanger TheFanger.com | InADazeFarms.com

This the first stab at  making a particle system that interacts with other instances of the same page which happen to be open.  
The idea is the particles will flow from one window to another regardless of screen positioning .
I saw a video on the internet i dont know where anymore of somebody who had two windows interacting.

I was hoping to use local storage to  make it all happen.

As of 2-16-23 I cant get it to work. This is where i stand as of 2-16-23 :D